+++ README.md
@@ -3,10 +3,9 @@
 
 [![Deploy](https://github.com/Nexters/tidify-server/actions/workflows/main.yml/badge.svg)](https://github.com/Nexters/tidify-server/actions/workflows/main.yml)
 
-## dep
-```
-pip install fastapi 'uvicorn[standard]' SQLAlchemy alembic PyJWT psycopg2-binary pytest-asyncio
-```
+## swagger
+
+http://tidify.herokuapp.com/docs
 
 ## dev
 ```bash
