+++ src/app/api/v1/tags.py
@@ -5,8 +5,8 @@ from fastapi_pagination.bases import AbstractPage
 from sqlalchemy.orm import Session
 
 from app.crud import tag_crud
-from app.models.models.tags import TagCreateRequest, TagDetailResponse, TagUpdateRequest
-from app.models.models.users import UserMe
+from app.models.tags import TagCreateRequest, TagDetailResponse, TagUpdateRequest
+from app.models.users import UserMe
 from app.services.user_svc import get_current_user
 from core.errors import exceptions
 from core.utils.query_utils import to_dict
