+++ src/app/api/v1_api.py
@@ -13,4 +13,4 @@ v1_router.include_router(bookmark_router, tags=["bookmarks"])
 v1_router.include_router(tag_router, tags=["tags"])
 v1_router.include_router(folder_router, tags=["folders"])
 v1_router.include_router(user_router, tags=["users"])
-v1_router.include_router(auth_router, tags=["auth"])
+v1_router.include_router(auth_router, tags=["oauth"])
