--- src/app/models/base.py
@@ -1,10 +1,0 @@
-from pydantic import BaseModel
-
-
-class OrmModel(BaseModel):
-    class Config:
-        orm_mode = True
-
-
-class CommandResponse(BaseModel):
-    id: int
