+++ src/app/services/user_svc.py
@@ -2,7 +2,7 @@ from fastapi import Depends
 from sqlalchemy.orm import Session
 
 from app.crud import user_crud
-from app.models.models.users import UserInput, UserToken, UserMe
+from app.models.users import UserInput, UserToken, UserMe
 from app.services.auth_svc import decode_token, AUTH_HEADER
 from core.errors.exceptions import UserNotFoundException
 from core.utils.query_utils import to_dict
