+++ src/tests/conftest.py
@@ -8,7 +8,7 @@ from pydantic import typing
 from sqlalchemy.orm import Session
 
 from app.crud import user_crud
-from app.models.models.users import UserInput, UserToken
+from app.models.users import UserInput, UserToken
 from app.services.auth_svc import create_access_token
 from core.consts import Phase, JWT_HEADER_NAME
 from database.conn import Base, db
